Deletados player.py e enemy.py //// jogo.py renomeado para main.py ////  Criado personagens.py, eh aqui que estarao as classes player e enemy
